pyproject.toml
[project]
name = "store-api-ux"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "pycryptodome>=3.23.0",
    "pyjwt>=2.10.1",
    "pyotp>=2.9.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
]
